requirements.txt
aiogram==3.24.0
python-dotenv==1.2.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-freezegun>=0.4.2
pytest-asyncio>=0.21.0
